refactor: add semantic versioning and /api prefix to routes
!BREAKING CHANGE. this follows best practices per stackoverflow blog here: https://stackoverflow.blog/2021/10/06/best-practices-for-authentication-and-authorization-for-rest-apis/. there are several ways to implement this, but i've opted just for URI versioning, as i don't think i'll be doing too much to cause too many breaking changes (see blog post here: https://dev.to/backplane/api-versioning-with-nodejs-express-pdd). this commit itself is a breaking change though, but only in development (nothing is deployed yet). api needs to be called with /api/v1/... rather than just /... straightaway.
