+++ src/app.ts
@@ -1,8 +1,6 @@
 import express from "express";
-import authRouter from "./routes/auth";
-import postsRouter from "./routes/posts";
-import usersRouter from "./routes/users";
 import "dotenv/config";
+import router from "./routes";
 
 if (!process.env.SECRET) {
 	throw new Error("SECRET variable missing in dotenv file.");
@@ -13,9 +11,7 @@ const app = express();
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 
-app.use("/", authRouter);
-app.use("/users", usersRouter);
-app.use("/posts", postsRouter);
+app.use("/api", router);
 
 const PORT = 6969;
 app.listen(PORT, (err) => {
