+++ src/routes/v1/posts.ts
@@ -9,7 +9,7 @@ import {
 	unpublishPost,
 } from "@/controllers/posts";
 import { attachPost } from "@/middleware/attach";
-import { authenticateWithJwt, checkIsAuthor } from "../middleware/auth";
+import { authenticateWithJwt, checkIsAuthor } from "@/middleware/auth";
 import commentsRouter from "./comments";
 
 const postsRouter = Router();
